fix(infra): remove node package strange dependencies and move to shared
