+++ apps/api/src/app/events/events.controller.ts
@@ -1,8 +1,7 @@
 import { Body, Controller, Delete, Param, Post, Scope, UseGuards } from '@nestjs/common';
 import { ApiOkResponse, ApiExcludeEndpoint, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { v4 as uuidv4 } from 'uuid';
-import { IJwtPayload, ISubscribersDefine } from '@novu/shared';
-import { TriggerRecipientSubscriber } from '@novu/node';
+import { IJwtPayload, ISubscribersDefine, TriggerRecipientSubscriber } from '@novu/shared';
 import { EventsPerformanceService, SendTestEmail, SendTestEmailCommand } from '@novu/application-generic';
 
 import {
