+++ apps/api/src/app/events/usecases/parse-event-request/parse-event-request.command.ts
@@ -1,6 +1,5 @@
 import { IsDefined, IsString, IsOptional } from 'class-validator';
-import { TriggerRecipients, TriggerRecipientSubscriber } from '@novu/node';
-import { ISubscribersDefine } from '@novu/shared';
+import { ISubscribersDefine, TriggerRecipients, TriggerRecipientSubscriber } from '@novu/shared';
 
 import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
