+++ apps/worker/src/app/workflow/usecases/message-matcher/types.ts
@@ -1,4 +1,4 @@
-import type { ITriggerPayload } from '@novu/node';
+import type { ITriggerPayload } from '@novu/shared';
 import type { SubscriberEntity } from '@novu/dal';
 
 export interface IFilterVariables {
