+++ libs/shared/src/dto/events/event.interface.ts
@@ -1,4 +1,9 @@
-export enum TriggerRecipientsTypeEnum {
-  SUBSCRIBER = 'Subscriber',
-  TOPIC = 'Topic',
-}
+import { ISubscribersDefine, ITopic, TopicKey } from '../../types';
+
+export type TriggerRecipientSubscriber = string | ISubscribersDefine;
+
+export type TriggerRecipient = TriggerRecipientSubscriber | ITopic;
+
+export type TriggerRecipients = TriggerRecipient[];
+
+export type TriggerRecipientsPayload = TriggerRecipientSubscriber | TriggerRecipients;
