+++ libs/shared/src/entities/subscriber/subscriber.interface.ts
@@ -1,4 +1,5 @@
 import { ChatProviderIdEnum, PushProviderIdEnum } from '../../consts';
+import { SubscriberCustomData } from '../../types';
 
 export interface ISubscriber {
   _id?: string;
@@ -29,5 +30,3 @@ export interface IChannelCredentials {
   webhookUrl?: string;
   deviceTokens?: string[];
 }
-
-export type SubscriberCustomData = { [key: string]: string | string[] | boolean | number | undefined };
