+++ libs/shared/src/types/index.ts
@@ -1,6 +1,7 @@
 export * from './builder';
 export * from './channel';
 export * from './environment';
+export * from './events';
 export * from './feature-flags';
 export * from './layout';
 export * from './message-template';
