+++ libs/testing/package.json
@@ -23,7 +23,6 @@
   "dependencies": {
     "@faker-js/faker": "^6.0.0",
     "@novu/dal": "^0.16.0",
-    "@novu/node": "^0.16.0",
     "@novu/shared": "^0.16.0",
     "JSONStream": "^1.3.5",
     "async": "^3.2.0",
