+++ packages/node/src/lib/novu.interface.ts
@@ -1,15 +1,7 @@
 import { AxiosInstance } from 'axios';
-import { ChannelTypeEnum } from '@novu/shared';
 
 export interface INovuConfiguration {
   backendUrl?: string;
-}
-
-export interface IAttachmentOptions {
-  mime: string;
-  file: Buffer;
-  name?: string;
-  channels?: ChannelTypeEnum[];
 }
 
 export class WithHttp {
